Change the way for adding log
Retire option -s and -d, use -m instead. Anything
above the first empty line of the argument of -m
is the subject, anything else below is the log
content (the data). When no -m option provided,
enter interactive mode, but will fail if the stdin
is not a tty, that is, a pipe, a file.